+++ spec/lib/ckeditor5/rails/engine_spec.rb
@@ -105,15 +105,29 @@ RSpec.describe CKEditor5::Rails::Engine do
 
     describe 'importmap initializer' do
       context 'when Importmap is defined' do
-        module Importmap # rubocop:disable Lint/ConstantDefinitionInBlock
-          module ImportmapTagsHelper; end
+        before do
+          Object.send(:remove_const, :Importmap) if defined?(Importmap)
+
+          module Importmap # rubocop:disable Lint/ConstantDefinitionInBlock
+            module ImportmapTagsHelper; end
+
+            class TestClass
+              include ImportmapTagsHelper
+            end
+          end
+        end
+
+        after do
+          Object.send(:remove_const, :Importmap) if defined?(Importmap)
         end
 
         it 'prepends CKEditor5 importmap helper' do
           initializer = described_class.initializers.find { |i| i.name == 'ckeditor5.importmap' }
           initializer.run(Rails.application)
 
-          expect(Importmap::ImportmapTagsHelper.ancestors)
+          Importmap::TestClass.new
+
+          expect(Importmap::TestClass.ancestors)
             .to include(CKEditor5::Rails::Hooks::Importmap::ImportmapTagsHelper)
         end
       end